refactor(package): 패키지 정리
