--- src/main/kotlin/domain/GameInfo.kt
@@ -1,6 +1,0 @@
-package domain
-
-import model.BetInfos
-import model.Participants
-
-class GameInfo(val participants: Participants, val playersInfo: BetInfos)
