+++ src/main/kotlin/domain/util/BetInfosBuilder.kt
@@ -1,0 +1,14 @@
+package domain.util
+
+import model.participants.Bet
+import model.participants.BetInfo
+import model.participants.Player
+
+class BetInfosBuilder {
+    private var infos = mutableMapOf<Player, Bet>()
+    fun bet(player: Player, bet: Bet) {
+        infos[player] = bet
+    }
+
+    fun build(): BetInfo = BetInfo(infos)
+}
