--- src/main/kotlin/model/BetInfos.kt
@@ -1,5 +1,0 @@
-package model
-
-data class BetInfos(val infos: Map<Player, Bet>) {
-    operator fun get(player: Player): Bet? = infos[player]
-}
