+++ src/main/kotlin/model/cards/Card.kt
@@ -1,3 +1,3 @@
-package model
+package model.cards
 
 data class Card(val rank: Rank, val suit: Suit)
