+++ src/main/kotlin/model/cards/Suit.kt
@@ -1,4 +1,4 @@
-package model
+package model.cards
 
 enum class Suit {
     DIAMOND, CLOVER, SPADE, HEART
