+++ src/main/kotlin/model/participants/BetInfo.kt
@@ -1,0 +1,5 @@
+package model.participants
+
+data class BetInfo(val info: Map<Player, Bet>) {
+    operator fun get(player: Player): Bet? = info[player]
+}
