+++ src/main/kotlin/model/participants/Dealer.kt
@@ -1,4 +1,6 @@
-package model
+package model.participants
+
+import model.cards.Hand
 
 class Dealer(hand: Hand, name: Name = Name(DEALER)) : Participant(hand, name) {
     override fun isHit(): Boolean = hand.sum() < HIT_STANDARD_POINT
