+++ src/main/kotlin/model/participants/Participants.kt
@@ -1,4 +1,4 @@
-package model
+package model.participants
 data class Participants(val dealer: Dealer, val players: Players) {
 
     fun toList(): List<Participant> = listOf(dealer as Participant) + players.toList() as List<Participant>
