+++ src/main/kotlin/model/participants/Players.kt
@@ -1,4 +1,4 @@
-package model
+package model.participants
 
 data class Players(private val players: List<Player>) {
     fun toList(): List<Player> = players.toList()
