+++ src/main/kotlin/model/result/GameResult.kt
@@ -1,13 +1,17 @@
-package model
+package model.result
+
+import model.participants.BetInfo
+import model.participants.Dealer
+import model.participants.Name
 
 class GameResult private constructor(val playersFinalResult: Map<Name, Int>) {
     fun getDealerProfitResult(): Int = playersFinalResult.values.sum() * -1
 
     companion object {
-        fun of(dealer: Dealer, betInfos: BetInfos): GameResult =
+        fun of(dealer: Dealer, betInfos: BetInfo): GameResult =
             GameResult(
                 buildMap {
-                    betInfos.infos.forEach {
+                    betInfos.info.forEach {
                         put(it.key.name, (it.value.amount * it.key.judge(dealer).rate).toInt())
                     }
                 }
