+++ src/test/kotlin/blackjack/model/BetInfosTest.kt
@@ -1,10 +1,10 @@
 package blackjack.model
 
-import model.Bet
-import model.BetInfos
-import model.Hand
-import model.Name
-import model.Player
+import model.cards.Hand
+import model.participants.Bet
+import model.participants.BetInfo
+import model.participants.Name
+import model.participants.Player
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
@@ -13,7 +13,7 @@ class BetInfosTest {
     fun `플레이어와 배팅금액을 연결지을 수 있다`() {
         val player = Player(Hand(emptyList()), Name("hyem"))
         val bet = Bet(10000)
-        val infos = BetInfos(mapOf(player to bet))
+        val infos = BetInfo(mapOf(player to bet))
         assertThat(infos[player]).isEqualTo(bet)
     }
 }
