+++ src/test/kotlin/blackjack/model/HandTest.kt
@@ -1,9 +1,9 @@
 package blackjack.model
 
-import model.Card
-import model.Hand
-import model.Rank
-import model.Suit
+import model.cards.Card
+import model.cards.Hand
+import model.cards.Rank
+import model.cards.Suit
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
