+++ src/test/kotlin/blackjack/model/ParticipantsTest.kt
@@ -1,11 +1,11 @@
 package blackjack.model
 
-import model.Dealer
-import model.Hand
-import model.Name
-import model.Participants
-import model.Player
-import model.Players
+import model.cards.Hand
+import model.participants.Dealer
+import model.participants.Name
+import model.participants.Participants
+import model.participants.Player
+import model.participants.Players
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
